+++ src/background.js
@@ -1,18 +1,18 @@
 import Dexie from 'dexie'
 import moment from 'moment';
-import {v4 as uuidv4} from 'uuid';
+import {v4 as UUIDv4} from 'uuid';
 
 const allPorts = [];
 
 let events = [];
 let socket, authToken, dontRetry = null;
-let settings = localStorage.getItem('settings')
-    ? JSON.parse(localStorage.getItem('settings'))
-    : {notify_online: false, notify_online_favorited: false, notify_notifications: false};
+let settings = localStorage.getItem('settings') ? JSON.parse(localStorage.getItem('settings')) : {
+    notify_online: false,
+    notify_online_favorited: false,
+    notify_notifications: false
+};
 
-let favorite_friends = localStorage.getItem('favorite_friends')
-    ? JSON.parse(localStorage.getItem('favorite_friends'))
-    : [];
+let favorite_friends = localStorage.getItem('favorite_friends') ? JSON.parse(localStorage.getItem('favorite_friends')) : [];
 
 const events_db = new Dexie('events_db');
 events_db.version(2).stores({
@@ -22,9 +22,9 @@ events_db.version(2).stores({
 events_db.events
     .where('date')
     .below(moment().subtract(24, 'hour').toDate())
-    .toArray()
-    .then(data => {
-        events_db.events.bulkDelete(data.map(e => e.id));
+    .delete()
+    .then((amount) => {
+        console.log(`%cDeleted ${amount} old events.`, 'color: #90c7ee; font-size: 20px;');
     });
 
 events_db.events
@@ -33,8 +33,7 @@ events_db.events
     .toArray()
     .then(data => {
         data.forEach(event => {
-            if (!event.uid)
-                event.uid = uuidv4();
+            if (!event.uid) event.uid = UUIDv4();
         });
 
         events = events.concat(data);
@@ -43,13 +42,11 @@ events_db.events
 function init() {
     if (browser.cookies) {
         browser.cookies.get({
-            name: 'auth',
-            url: 'https://vrchat.com'
+            name: 'auth', url: 'https://vrchat.com'
         }).then(useCookie);
     } else if (chrome.cookies) {
         chrome.cookies.get({
-            name: 'auth',
-            url: 'https://vrchat.com'
+            name: 'auth', url: 'https://vrchat.com'
         }, useCookie);
     }
 }
@@ -64,9 +61,15 @@ function useCookie(cookie) {
         authToken = cookie.value;
 
         sendMessageToPorts({type: 'all_events', events});
-        createSocket(`wss://pipeline.vrchat.cloud/?authToken=${cookie.value}`);
+        createSocket(`wss://vrchat.com/?authToken=${cookie.value}`);
     }
 }
+
+chrome.notifications.onClicked.addListener(function (notificationId) {
+    if (notificationId === 'openLoginPage') {
+        chrome.tabs.create({url: 'https://vrchat.com/login'});
+    }
+});
 
 function createSocket(url) {
     if (socket) socket.close(1000, 'known_close');
@@ -74,18 +77,56 @@ function createSocket(url) {
     socket.onmessage = handleVRCEvent;
 
     socket.onclose = (ev) => {
-        if (ev.reason !== 'known_close')
-            setTimeout(() => {
-                if (!dontRetry)
-                    createSocket(url);
-                dontRetry = false;
-            }, 2000);
+        console.log('%cSocket Closed!', 'color: #ee9090; font-size: 50px;');
+        console.log(`Code: ${ev.code}, Reason: "${ev.reason}"`);
+
+        fetch('https://vrchat.com/api/1/auth/user')
+            .then(response => {
+                if (response.status === 503) console.log('%Cloudflare error?!', 'color: #eeba90; font-size: 50px;'); else return response.json();
+            })
+            .then(data => {
+                if (data.error && data.error.status_code === 401) setExtensionStatus(false);
+            });
+
+        if (ev.reason !== 'known_close') setTimeout(() => {
+            if (!dontRetry) createSocket(url);
+            dontRetry = false;
+        }, 2000);
     };
 
     socket.onerror = (ev) => console.error(ev);
 
     socket.onopen = () => {
         console.log('%cSocket Initiated!', 'color: #90ee90; font-size: 50px;');
+
+        setExtensionStatus(true);
+    }
+}
+
+function setExtensionStatus(online = true) {
+    if (online) {
+        chrome.browserAction.setIcon({
+            path: {
+                '19': 'icons/vrce-logo-19_x_19.png', '38': 'icons/vrce-logo-38_x_38.png'
+            }
+        });
+
+        chrome.browserAction.setTitle({title: 'VRCe - Online'});
+    } else {
+        chrome.notifications.create('openLoginPage', {
+            title: 'Disconnected',
+            message: 'Click to open login page',
+            type: 'basic',
+            iconUrl: 'icons/vrce-logo-128_x_128.png'
+        });
+
+        chrome.browserAction.setIcon({
+            path: {
+                '19': 'icons/vrce-logo-19_x_19-offline.png', '38': 'icons/vrce-logo-38_x_38-offline.png'
+            }
+        });
+
+        chrome.browserAction.setTitle({title: 'VRCe - Offline'});
     }
 }
 
@@ -98,26 +139,21 @@ function handleVRCEvent(ev) {
         return;
     }
 
-    event.uid = uuidv4();
+    event.uid = UUIDv4();
     event.date = new Date();
     event.content = parseContent(event.content);
 
     events.push(event);
     events_db.events.add(event);
 
-    if (
-        (settings.notify_online && event.type === 'friend-online')
-        || (settings.notify_online_favorited && event.type === 'friend-online' && favorite_friends.includes(event.content.user.id))
-    )
-        chrome.notifications.create({
-            title: 'Friend Online',
-            message: event.content.user.displayName,
-            type: 'basic',
-            iconUrl: 'icons/vrce-logo-128_x_128.png'
-        });
+    if ((settings.notify_online && event.type === 'friend-online') || (settings.notify_online_favorited && event.type === 'friend-online' && favorite_friends.includes(event.content.user.id))) chrome.notifications.create({
+        title: 'Friend Online',
+        message: event.content.user.displayName,
+        type: 'basic',
+        iconUrl: 'icons/vrce-logo-128_x_128.png'
+    });
 
-    if (settings.notify_notifications && event.type === 'notification')
-        notifyNotification(event).then();
+    if (settings.notify_notifications && event.type === 'notification') notifyNotification(event).then();
 
     sendMessageToPorts({type: 'new_events', event});
 }
@@ -169,6 +205,7 @@ async function notifyNotificationOptions(event) {
             notification.contextMessage = event.content.details.inviteMessage;
             break;
         case 'inviteResponse':
+        case 'requestInviteResponse':
             notification.title = 'Reply';
             notification.message = `${event.content.senderUsername} replied to you`;
             notification.iconUrl = 'icons/reply.png';
@@ -201,9 +238,14 @@ async function notifyNotificationOptions(event) {
     } else if (port.name === 'popup-settings') {
         port.postMessage({type: 'settings', settings});
 
-        port.onMessage.addListener(settingsMsg => {
-            settings = settingsMsg;
-            saveSettings();
+        port.onMessage.addListener(msg => {
+            if (msg.type === 'update_settings') {
+                settings = msg.settings;
+                saveSettings();
+            } else if (msg.type === 'clear_events') {
+                events = [];
+                events_db.events.clear();
+            }
         });
     } else if (port.name === 'popup-app') {
         init();
@@ -215,12 +257,13 @@ async function notifyNotificationOptions(event) {
                 const userId = msg.user_id;
                 const index = favorite_friends.indexOf(userId);
 
-                if (index === -1) favorite_friends.push(userId);
-                else favorite_friends.splice(index, 1);
+                if (index === -1) favorite_friends.push(userId); else favorite_friends.splice(index, 1);
 
                 saveFavoriteFriends();
 
                 port.postMessage({type: 'favorite_friends', favorite_friends});
+            } else if (msg.type === 'logout') {
+                if (socket) socket.close(1000, 'known_close');
             }
         });
     }